move provisioning-info fsm test to patch_shoot_test
